Add internal state in region selectio input
