--- src/root/components/ThreeWForm/RegionSelectionInput/styles.module.scss
@@ -1,67 +1,0 @@
-@import '~styles/settings/variables';
-
-.modal {
-  display: flex;
-  flex-grow: 1;
-  min-height: 70vh;
-  padding: $spacing*3;
-  gap: $spacing*2;
-
-  .modal-body {
-    padding: 0;
-  }
-
-  .modal-content {
-    padding: 0;
-
-    .secondary-header {
-      padding: 0;
-    }
-
-    .main-content {
-      display: flex;
-
-      .map-container {
-        flex-grow: 1;
-        outline: none;
-        display: flex;
-        flex-direction: column;
-        position: relative;
-
-        .map-content {
-          width: 100%;
-          height: 70vh;
-        }
-
-        .map-disclaimer {
-          position: absolute;
-          left: $mapbox-icon-width + $spacing*0.5;
-          bottom: $spacing*0.25;
-        }
-
-        :global {
-          .mapboxgl-popup {
-            height: 18rem;
-          }
-        }
-
-        .wrapper-for-sidebar {
-          display: flex;
-          flex-grow: 1;
-          position: relative;
-
-          .map {
-            flex-grow: 1;
-          }
-        }
-      }
-
-      .side-bar {
-        width: calc(8rem + 18%);
-        padding: $spacing;
-        overflow: auto;
-        height: 70vh;
-      }
-    }
-  }
-}
